+++ src/index.tsx
@@ -10,9 +10,11 @@ import * as Constants from "models/Constants.ts";
 import "assets/css/material-dashboard-react.css?v=1.6.0";
 
 // core components
-import Admin from "layouts/Admin";
+import Panel from "layouts/Panel.tsx";
 import Main from "layouts/Main.tsx";
 import WaitingComponent from "hocs/WaitingComponent";
+import { ThemeProvider } from "@material-ui/core/styles";
+import theme from "assets/theme";
 
 const hist = createBrowserHistory();
 
@@ -20,15 +22,17 @@ App.setHistory(hist);
 App.start();
 
 ReactDOM.render(
-  <Router history={hist}>
-    <Switch>
-      <Route path={Constants.ADMIN_ROUTE} component={WaitingComponent(Admin)}/>
-      <Route path={Constants.LOGIN_ROUTE} component={WaitingComponent(Main)}/>
-      <Route path={Constants.SIGN_UP_ROUTE} component={WaitingComponent(Main)}/>
-      <Route path={Constants.ROOT_ROUTE} component={WaitingComponent(Main)}/>
-      <Redirect to={Constants.START_PAGE_ROUTE}/>
-      <Route path="*" component={WaitingComponent(Admin)}/>
-    </Switch>
-  </Router>,
+  <ThemeProvider theme={theme}>
+    <Router history={hist}>
+      <Switch>
+        <Route path={Constants.PANEL_ROUTE} component={WaitingComponent(Panel)}/>
+        <Route path={Constants.LOGIN_ROUTE} component={WaitingComponent(Main)}/>
+        <Route path={Constants.SIGN_UP_ROUTE} component={WaitingComponent(Main)}/>
+        <Route path={Constants.ROOT_ROUTE} component={WaitingComponent(Main)}/>
+        <Redirect to={Constants.START_PAGE_ROUTE}/>
+        <Route path="*" component={WaitingComponent(Panel)}/>
+      </Switch>
+    </Router>
+  </ThemeProvider>,
   document.getElementById("root")
 );