+++ src/models/App.ts
@@ -1,12 +1,16 @@
-import { action, computed, IObservableArray, IReactionDisposer, observable, reaction, when } from "mobx";
+import { action, computed, IReactionDisposer, observable, reaction, when } from "mobx";
 import { History } from "history";
+
+// interfaces
 import { IFlow } from "interfaces/IFlow";
+import { IUser } from "interfaces/IUser";
+import { IRole } from "interfaces/IRole";
+
+// models
 import { RoleStore } from "models/Role/RoleStore.ts";
 import { UserStore } from "models/User/UserStore.ts";
 import { Auth } from "models/Auth/Auth.ts";
 import * as Constants from "models/Constants.ts";
-import { IUser } from "interfaces/IUser";
-import { IRole } from "interfaces/IRole";
 import { Apps } from "models/App/AppsStore";
 import { Settings } from "models/Settings";
 import { AppDataStore } from "models/App/AppDataStore";
@@ -73,11 +77,12 @@ export class AppStore implements IFlow {
     if (this.user.anonymous) {
       this.navigationHistory.push(Constants.ROOT_ROUTE);
     } else {
-      console.log(2323232323, this.navigationHistory.location.pathname, this.navigationHistory.location.pathname.includes(Constants.ADMIN_ROUTE));
-      if (!this.navigationHistory.location.pathname.includes(Constants.ADMIN_ROUTE)) {
-        this.navigationHistory.push(Constants.EVENTS_USERS_LIST_ROUTE);
+      console.log(2323232323, this.navigationHistory.location.pathname, this.navigationHistory.location.pathname.includes(Constants.PANEL_ROUTE));
+      if (!this.navigationHistory.location.pathname.includes(Constants.PANEL_ROUTE)) {
+        this.navigationHistory.push(Constants.DASHBOARD_ROUTE);
       } else {
-        this.navigationHistory.replace(this.navigationHistory.location.pathname);
+        // causes extra screen reloading
+        // this.navigationHistory.replace(this.navigationHistory.location.pathname);
       }
     }