+++ src/utils/convertDate.ts
@@ -14,3 +14,12 @@ export default function convertDate(object: any) {
     }
   });
 }
+
+export function dateToString(date?: Date) {
+  date = date || new Date();
+  let year = date.getFullYear();
+  let month = (1 + date.getMonth()).toString().padStart(2, '0');
+  let day = date.getDate().toString().padStart(2, '0');
+
+  return month + '/' + day + '/' + year;
+}