+++ src/views/UserProfile/UserProfile.jsx
@@ -1,7 +1,9 @@
 import React from "react";
+
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
 import InputLabel from "@material-ui/core/InputLabel";
+
 // core components
 import GridItem from "components/Grid/GridItem.tsx";
 import GridContainer from "components/Grid/GridContainer.tsx";